Enforced Google's specification regarding back-off mode
As explained in the specs
(https://developers.google.com/safe-browsing/v4/request-frequency), all
non-successful return code shall trigger a back-off mode for the API. It
was previously implemented for the threatListUpdates.fetch method, but
not for fullHashes.find.
For that purpose, back-off handling has been moved from the database
scope to the API scope, providing a unified interface and behavior
regardless of the used method.